Add basic winning computer logic
Add basic logic to the computer move so that it will first check to see if there are two of its symbol in any row, column or diagonal and if there is also a free space, then it will add its symbol to this free space, winning the game. Otherwise, it will add its symbol to the first free space available.